core: move early_ta implementation to embedded_ts
Ealy_ta's are similar to embedded SPs. Move all shared logic to the
embedded_ts.

Acked-by: Jerome Forissier <jerome@forissier.org>
